+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next
 
-- fix: getFeatureFlagPayload wasn't sending the $feature_flag_called event ([#185](https://github.com/PostHog/posthog-android/pull/185))
+- fix: getFeatureFlagPayload wasn't sending the $feature_flag_called event ([#226](https://github.com/PostHog/posthog-android/pull/226))
 
 ## 3.11.2 - 2025-02-04