+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Next
 
+## 3.11.3 - 2025-02-25
+
+- feat: support quota limiting for feature flags ([#228](https://github.com/PostHog/posthog-android/pull/228))
+
 ## 3.11.2 - 2025-02-04
 
 - fix: touches fall back to single touches if out of bounds ([#221](https://github.com/PostHog/posthog-android/pull/221))